Add lane-local cooperative-matrix lowering foundations (#1825)
* Infer cooperative-matrix expression contracts

Track lexical value types while parsing and infer cooperative-matrix result
contracts for elementwise, multiply, and multiply-accumulate operations.
Apply declaration, assignment, and return constraints while preserving scalar
element expression types and lexical shadowing.

* Add DirectX lane-local matrix fragments

Provide an explicit opt-in representation for registered cooperative-matrix
fragment mappings and lower element, negate, and elementwise operations to
lane-local HLSL vectors. Preserve the default fail-closed path and reject
incomplete, incompatible, or non-lane-local contracts with diagnostics.

* Add OpenGL lane-local matrix fragments

Provide an explicit opt-in GLSL fragment struct and helpers for registered
cooperative-matrix mappings. Lower element, negate, and elementwise operations,
retain structured fail-closed diagnostics for unsupported contracts, and guard
unresolved private-pointer address bases during range analysis.

* Record the cooperative-matrix target frontier

Update the pinned MLX evidence after shared expression inference and the
opt-in DirectX and OpenGL lane-local lowering foundations. Record the next
private-pointer boundaries without claiming complete translation, runtime
integration, or numerical parity.

* Refresh cooperative-matrix support evidence

Regenerate the support inventory after adding the DirectX and OpenGL
cooperative-matrix lowering tests and diagnostics.

* Preserve cooperative-matrix target diagnostics

Keep invalid cooperative-matrix operations in permissively parsed function
bodies so target generators can report their structured diagnostics. Retain
strict parser rejection and require an explicit Vulkan accumulator role when
shared inference produces a role-neutral result.

* Keep the const-reference fixture validator-neutral

Use ordinary integer indices in the nested const-reference alias fixture so
the HLSL fallback validator exercises the intended alias lowering without
requiring native 16-bit HLSL support.

+++ tests/test_backend/test_metal/test_codegen.py
@@ -7283,14 +7283,14 @@ def test_codegen_nested_const_reference_alias_reaches_native_targets(tmp_path):
     using namespace metal;
 
     struct Tile {
-        static constexpr short width = 2;
+        static constexpr int width = 2;
         float2 val_frags[4];
 
-        constexpr thread float2& frag_at(short i, short j) {
+        constexpr thread float2& frag_at(int i, int j) {
             return val_frags[i * width + j];
         }
 
-        constexpr const thread float2& frag_at(short i, short j) const {
+        constexpr const thread float2& frag_at(int i, int j) const {
             return val_frags[i * width + j];
         }
     };
@@ -7298,9 +7298,9 @@ def test_codegen_nested_const_reference_alias_reaches_native_targets(tmp_path):
     struct StoreLoop {
         Tile Ctile;
 
-        void store(thread float2& stored, short i, short j) const {
+        void store(thread float2& stored, int i, int j) const {
             thread const auto& accum = Ctile.frag_at(i, j);
-            for (short k = 0; k < 2; ++k) {
+            for (int k = 0; k < 2; ++k) {
                 stored[k] = accum[k];
             }
         }
